api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.6",
    "greenlet>=3.0.0",
    "fiona>=1.10.1",
    "geoalchemy2>=0.18.3",
    "geopandas>=1.1.2",
    "bcrypt>=4.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.47",
    "fastapi-mcp>=0.4.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "reportlab>=4.2.0",
    "matplotlib>=3.9.0",
]

[project.scripts]
start = "api.main:start"

